+++ registry/components/table.py
@@ -97,11 +97,13 @@ def TableRow(
 
 def TableHead(
     *children: Any,
+    scope: str = "col",
     cls: str = "",
     **kwargs: Any,
 ) -> FT:
     return Th(
         *children,
+        scope=scope,
         data_slot="table-head",
         cls=cn(
             "text-foreground h-10 px-2 text-left align-middle font-medium whitespace-nowrap [&:has([role=checkbox])]:pr-0 [&>[role=checkbox]]:translate-y-[2px]",